+++ priv/repo/migrations/20211021005023_create_entities.exs
@@ -6,7 +6,7 @@ defmodule Thexr.Repo.Migrations.CreateEntities do
       add :id, :binary_id, primary_key: true
       add :name, :string, null: false
       add :type, :string, null: false
-      add :space_id, references(:spaces, on_delete: :delete_all, type: :binary_id)
+      add :space_id, references(:spaces, on_delete: :delete_all, type: :binary_id), null: false
       add :parent_id, references(:entities, on_delete: :delete_all, type: :binary_id)
       # left
       # right
@@ -16,6 +16,6 @@ defmodule Thexr.Repo.Migrations.CreateEntities do
 
     create index(:entities, [:space_id])
     create index(:entities, [:parent_id])
-    create unique_index(:entities, [:space_id, :name])
+    # create unique_index(:entities, [:space_id, :name])
   end
 end