+++ Changelog.md
@@ -3,3 +3,7 @@
 ## [0.0.0] - 2021-10-20
 ### Added
 - Spaces, Entities
+## [0.0.1] - 2021-10-21
+<!-- - Space has one or more 'scene' entities. -->
+- Have a list of entities that can be added to a space
+- Box, Sphere, Cone